+++ chat21/src/main/java/org/chat21/android/utils/TimeUtils.java
@@ -1,6 +1,8 @@
 package org.chat21.android.utils;
 
-import org.ocpsoft.prettytime.PrettyTime;
+import android.content.Context;
+
+import org.chat21.android.R;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
@@ -12,119 +14,71 @@ import java.util.Locale;
  * Created by stefano on 05/09/2015.
  */
 public class TimeUtils {
-//    private static final String TAG = TimeUtils.class.getName();
 
-    private static Long ONE_DAY_IN_MILLIS = Long.valueOf(86400000);  // 1 day = 86400000 millis
+    public static boolean isDateToday(long milliSeconds) {
+        Calendar toCheck = Calendar.getInstance();
+        toCheck.setTimeInMillis(milliSeconds);
 
-    /**
-     * Convert a Long timestamp into a string
-     *
-     * @param timestamp the timestamp
-     * @param locale    the locale format
-     * @return the date string
-     */
-    public static String getFormattedDate(Long timestamp, Locale locale) {
-//        Log.d(TAG, "getFormattedDate");
+        Date toCheckDate = toCheck.getTime();
 
-        Long currentTime = getCurrentTime(locale);
-        Long timeDiff = currentTime - timestamp;
-        Long onedayInMillis = ONE_DAY_IN_MILLIS;
+        toCheck.setTimeInMillis(System.currentTimeMillis());
+        toCheck.set(Calendar.HOUR_OF_DAY, 0);
+        toCheck.set(Calendar.MINUTE, 0);
+        toCheck.set(Calendar.SECOND, 0);
 
-        /*
-         * if the difference between the current time and the timestamp is
-         * 1 day or more than show the compete timestamp else show the timestamp as
-         * HH:mm
-         */
-        DateFormat dateFormat;
-        if (timeDiff <= onedayInMillis) {
-            dateFormat = new SimpleDateFormat("HH:mm");
+        Date startDate = toCheck.getTime();
+
+        return toCheckDate.compareTo(startDate) > 0;
+    }
+
+    public static boolean isYesterday(long milliSeconds) {
+
+        Calendar toCheck = Calendar.getInstance();
+        toCheck.setTimeInMillis(milliSeconds); // your date
+
+        Calendar yesterday = Calendar.getInstance(); // today
+        yesterday.add(Calendar.DAY_OF_YEAR, -1); // yesterday
+
+        boolean isYesterday = yesterday.get(Calendar.YEAR) == toCheck.get(Calendar.YEAR)
+                && yesterday.get(Calendar.DAY_OF_YEAR) == toCheck.get(Calendar.DAY_OF_YEAR);
+
+        return isYesterday ? true : false;
+    }
+
+    public static boolean isDateInCurrentWeek(long milliSeconds) {
+        Calendar currentCalendar = Calendar.getInstance();
+        int week = currentCalendar.get(Calendar.WEEK_OF_YEAR);
+        int year = currentCalendar.get(Calendar.YEAR);
+
+        Calendar targetCalendar = Calendar.getInstance();
+        targetCalendar.setTimeInMillis(milliSeconds);
+        int targetWeek = targetCalendar.get(Calendar.WEEK_OF_YEAR);
+        int targetYear = targetCalendar.get(Calendar.YEAR);
+
+        return week == targetWeek && year == targetYear;
+    }
+
+    // https://www.ibm.com/support/knowledgecenter/en/SSMKHH_9.0.0/com.ibm.etools.mft.doc/ak05616_.htm
+    public static String getFormattedTimestamp(Context context, long milliSeconds) {
+        if (isDateToday(milliSeconds)) {
+            return formatTimestamp(milliSeconds, "HH:mm");
+        } else if (isYesterday(milliSeconds)) {
+            return context.getString(R.string.yesterday);
+        } else if (isDateInCurrentWeek(milliSeconds)) {
+            return formatTimestamp(milliSeconds, "EEEE");
         } else {
-            dateFormat = new SimpleDateFormat("dd-MM-yyyy HH:mm");
+            return formatTimestamp(milliSeconds, "dd/MM/yy");
         }
+    }
+
+    public static String formatTimestamp(Long timestamp, String pattern) {
+        DateFormat dateFormat = new SimpleDateFormat(pattern);
         return dateFormat.format(timestamp);
     }
 
-    // return the current time in millis
-    private static Long getCurrentTime(Locale locale) {
-//        Log.d(TAG, "getCurrentTime");
-        return Calendar.getInstance(locale).getTimeInMillis();
-    }
-
-    public static String fixPrettyTimeFutureMessage(String convertedTimestamp) {
-//        Log.d(TAG, "fixPrettyTimeFutureMessage");
-        String fixedTimestamp = convertedTimestamp;
-        if (convertedTimestamp.compareTo("fra poco") == 0) {
-            fixedTimestamp = "adesso";
-        }
-        return fixedTimestamp;
-    }
-
-    public static String getFormattedTimestamp(long timestampLong) {
-//        Log.d(TAG, "getFormattedTimestamp");
-        PrettyTime p = new PrettyTime();
-        String timestamp =
-                p.format(new Date((timestampLong)));
-        timestamp = TimeUtils.fixPrettyTimeFutureMessage(timestamp);
-        return timestamp;
-    }
-
-    // resolve Issue #38
-    public static String timestampToHour(long timestamp) {
-        Date date = new Date(timestamp);
-        SimpleDateFormat sdf = new SimpleDateFormat("HH:mm");
-        return sdf.format(date);
-    }
-
-    public static String timestampToStrDate(long timestamp) {
-        int currentYear = getCurrentYear();
-        int timestampYear = getYearFromTimestamp(timestamp);
-
-        // if it is the current year hide it, otherwise show it
-        String pattern = "dd MMMM yyyy";
-        if (timestampYear == currentYear) {
-            pattern = "dd MMMM";
-        }
-
-        // timestamp date
-        Date timestampDate = new Date(timestamp);
-        SimpleDateFormat sdf = new SimpleDateFormat(pattern);
-        return sdf.format(timestampDate);
-    }
-
-    // returns the date of week for the date
     public static String getDayOfWeek(Date date) {
         Calendar cal = Calendar.getInstance();
         cal.setTime(date);
         return cal.getDisplayName(Calendar.DAY_OF_WEEK, Calendar.LONG, Locale.getDefault());
-    }
-
-    // check if a timestamp is the current date
-    public static boolean isDateToday(long milliSeconds) {
-        Calendar calendar = Calendar.getInstance();
-        calendar.setTimeInMillis(milliSeconds);
-
-        Date getDate = calendar.getTime();
-
-        calendar.setTimeInMillis(System.currentTimeMillis());
-        calendar.set(Calendar.HOUR_OF_DAY, 0);
-        calendar.set(Calendar.MINUTE, 0);
-        calendar.set(Calendar.SECOND, 0);
-
-        Date startDate = calendar.getTime();
-
-        return getDate.compareTo(startDate) > 0;
-    }
-
-    // returns the year of the current date
-    private static int getCurrentYear() {
-        Calendar cal = Calendar.getInstance();
-        return cal.get(Calendar.YEAR);
-    }
-
-    // returns the year from the timestamp
-    private static int getYearFromTimestamp(long timestamp) {
-        Calendar cal = Calendar.getInstance();
-        cal.setTimeInMillis(timestamp);
-        return cal.get(Calendar.YEAR);
     }
 }